+++ src/icemac/songbeamer/__init__.py
@@ -2,4 +2,4 @@
 ENCODING = 'cp1252'
 
 # re-imports to ease usage
-from .sng import SNG  # noqa: F401
+from .sng import *  # noqa: F401,F403